+++ src/main/java/com/iexec/worker/pubsub/StompClientService.java
@@ -288,6 +288,8 @@ public class StompClientService {
         public void handleTransportError(StompSession session, Throwable exception) {
             log.error("STOMP transport error [session: {}, isConnected: {}, exception: {}]",
                     session.getSessionId(), session.isConnected(), exception.getMessage());
+            // notify subscribers
+            eventPublisher.publishEvent(new SessionLostEvent());
             requestNewSession();
         }
     }